Added color scenarios

+++ scripts/language/extract_champions.sh
@@ -25,12 +25,6 @@ then
   exit 2
 fi
 
-if ls $folder/champs/* > /dev/null 2>&1
-then
-  echo "Target folder already populated. Aborting"
-  exit 3
-fi
-
 mkdir -vp $folder/champs
 cd $folder/champs
 
@@ -38,7 +32,9 @@ ls -v ../gen*/lg*.dna | xargs jq -r '"\(input_filename) \(.alreadyEval)"' \
   | grep -v " true$" | cut -d ' ' -f1 | while read c
 do
   l=$(sed 's|.*/gen\(.*\)/lg.*.dna|\1|' <<< "$c"| xargs printf "gen%0*d.dna" $width)
-  ln -sv $c $l
+  [ ! -f $l ] && ln -sv $c $l
 done
 
-ln -sv $(ls *.dna | tail -n 1) last.dna
+last=$(ls gen*.dna | tail -n 1)
+[ -f last.dna ] && [ $(readlink -e $last) != $(readlink -e last.dna) ] && rm -v last.dna
+[ ! -f last.dna ] && ln -sv $last last.dna
